a little organization for the testing suite
