+++ tests/test_distributions.py
@@ -18,5 +18,5 @@ def test_simulate_signal_from_gaussian_distributions(
     scattering_theory = request.getfixturevalue(scattering_theory)
     instrument_config = request.getfixturevalue(instrument_config)
     image_model = cxs.ContrastImageModel(instrument_config, scattering_theory)
-    distribution = cls(image_model)
+    distribution = cls(image_model, normalizes_signal=False)
     np.testing.assert_allclose(image_model.render(), distribution.compute_signal())
