+++ tests/test_image_shape.py
