+++ tests/test_structural_ensemble.py
